removed source code from comments, added xxh3 reference
