+++ src/test/java/com/dynatrace/hash4j/hashing/Komihash4_3ReferenceData.java
@@ -68,94 +68,8 @@ final class Komihash4_3ReferenceData {
   }
 
   private static List<ReferenceRecord> createReferenceData() {
-    // the hashes are calculated using the c++ reference implementation on an Intel x86 architecture
-    // with little endianness
-    // the reference implementation was taken from
-    // https://github.com/avaneev/komihash/tree/e107760596dc5e883e26a58f81a5fd653061bd5a
-
-    /*
-    #include "komihash.h"
-
-    #include <iostream>
-    #include <iomanip>
-    #include <random>
-
-    using namespace std;
-
-    int main(int argc, char *argv[])
-    {
-
-      mt19937_64 rng(0);
-
-      uint64_t maxSize = 200;
-      uint64_t numExamplesPerSize = 10;
-
-      uniform_int_distribution<uint8_t> dist(0, 255);
-
-      for (uint64_t size = 0; size <= maxSize; ++size)
-      {
-        vector<uint8_t> data(size);
-        for (uint64_t i = 0; i < numExamplesPerSize; ++i)
-        {
-          for (uint64_t k = 0; k < size; ++k)
-          {
-            data[k] = dist(rng);
-          }
-          uint64_t seed = rng();
-
-          uint64_t hash0 = komihash((char *)(&data[0]), size, 0);
-          uint64_t hash1 = komihash((char *)(&data[0]), size, seed);
-
-          cout << "builder.add(0x";
-          cout << hex << setfill('0') << setw(16) << hash0;
-          cout << "L, 0x";
-          cout << hex << setfill('0') << setw(16) << hash1;
-          cout << "L, 0x";
-          cout << hex << setfill('0') << setw(16) << seed << 'L';
-          cout << ", \"";
-          for (uint64_t k = 0; k < size; ++k)
-            cout << hex << setfill('0') << setw(2) << static_cast<uint64_t>(data[k]);
-          cout << "\");";
-
-          cout << endl;
-        }
-      }
-
-      // some more special cases
-      {
-        uint64_t size = 64;
-        vector<uint8_t> data(size);
-        for (uint64_t s = 0; s < 256; ++s)
-        {
-          for (uint64_t i = 127; i <= 128; ++i)
-          {
-            for (uint64_t k = 0; k < size - 1; ++k)
-            {
-              data[k] = dist(rng);
-            }
-            data[63] = i;
-            uint64_t seed = rng();
-            seed = ((seed >> 8) << 8) | s;
-
-            uint64_t hash0 = komihash((char *)(&data[0]), size, 0);
-            uint64_t hash1 = komihash((char *)(&data[0]), size, seed);
-
-            cout << "builder.add(0x";
-            cout << hex << setfill('0') << setw(16) << hash0;
-            cout << "L, 0x";
-            cout << hex << setfill('0') << setw(16) << hash1;
-            cout << "L, 0x";
-            cout << hex << setfill('0') << setw(16) << seed << 'L';
-            cout << ", \"";
-            for (uint64_t k = 0; k < size; ++k)
-              cout << hex << setfill('0') << setw(2) << static_cast<uint64_t>(data[k]);
-            cout << "\");";
-
-            cout << endl;
-          }
-        }
-      }
-    }*/
+    // the reference data was computed using references/komihash_4_3/reference_data.cpp
+    // on an Intel x86 architecture with little endianness
 
     ReferenceDataBuilder builder = new ReferenceDataBuilder();
     builder.add(0xb7683ea7430132b4L, 0xb22efc38246bbc44L, 0x28e837c5cb41dc3eL, "");
