+++ src/test/java/com/dynatrace/hash4j/hashing/XXH3ReferenceData.java
@@ -68,55 +68,9 @@ final class XXH3ReferenceData {
   }
 
   private static List<ReferenceRecord> createReferenceData() {
-    // the hashes are calculated using the c++ reference implementation on an Intel x86 architecture
-    // with little endianness
-    // the reference implementation was taken from
-    // https://github.com/Cyan4973/xxHash/tree/d8aff149b5b9b9461f4b01fc5ab4842112d2c867
+    // the reference data was computed using references/xxh3/reference_data.cpp
+    // on an Intel x86 architecture with little endianness
 
-    /*
-      #define XXH_INLINE_ALL
-      #include "xxhash.h"
-      #include <iostream>
-      #include <iomanip>
-      #include <random>
-      using namespace std;
-      int main(int argc, char *argv[])
-      {
-          mt19937_64 rng(0);
-          uint64_t maxSize = 200;
-          uint64_t numExamplesPerSize = 10;
-          uint32_t defaultSeed = 0;
-          uniform_int_distribution<uint8_t> dist(0, 255);
-          for (uint64_t size = 0; size <= maxSize; ++size)
-          {
-              vector<uint8_t> data(size);
-              for (uint64_t i = 0; i < numExamplesPerSize; ++i)
-              {
-                  for (uint64_t k = 0; k < size; ++k)
-                  {
-                      data[k] = dist(rng);
-                  }
-                  uint64_t seed = rng();
-
-                  uint64_t hash0 = XXH3_64bits((char *)(&data[0]), size);
-                  uint64_t hash1 = XXH3_64bits_withSeed((char *)(&data[0]), size, seed);
-
-                  cout << "builder.add(0x";
-                  cout << hex << setfill('0') << setw(16) << hash0;
-                  cout << "L,0x";
-                  cout << hex << setfill('0') << setw(16) << hash1;
-                  cout << "L,0x";
-                  cout << hex << setfill('0') << setw(16) << seed << 'L';
-                  cout << ",\"";
-                  for (uint64_t k = 0; k < size; ++k)
-                      cout << hex << setfill('0') << setw(2) << static_cast<uint64_t>(data[k]);
-                  cout << "\");";
-
-                  cout << endl;
-              }
-          }
-      }
-    */
     ReferenceDataBuilder builder = new ReferenceDataBuilder();
     builder.add(0x2d06800538d394c2L, 0x98f6dd3cf2fcdd99L, 0x28e837c5cb41dc3eL, "");
     builder.add(0x2d06800538d394c2L, 0xd43b77e4a90dbbfdL, 0xfdfd3a7c3e40f98bL, "");
